First stab at a TLA+ implementation of Accord
Focused on the fast path consensus only. IOW this covers less than
half of algorithm 1. Also, actual tallying of votes, not implemented
yet. Focus here was just to learn how to pass state between coordinator
and data node.

Chose to use PlusCal for the spec. (This was probably a mistake.)

henrik