New attempt at a TLA+ model, fast path only
This implements a queue based approach, and I've been able to create
transactions and send them from coordinator to all nodes/shards and
back. 99% likelihood there are still errors, and in any case
checking for conflicts and dependencies is missing. This is
just to demonstrate that I can get messages through to a fast path
commit.